refactor : api 응답 유틸 메서드 리팩토링
