+++ be/src/main/java/com/sharetravel/global/api/ApiResponseCode.java
@@ -1,4 +1,4 @@
-package com.sharetravel.global;
+package com.sharetravel.global.api;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
